types changed: int to unsigned int

+++ bowling/inc/game.hpp
@@ -2,24 +2,24 @@
 
 #include <iostream>
 
-constexpr int MAX_PINS{10};
-constexpr int MAX_FRAMES{10};
-constexpr int MAX_ROLLS{21};
-constexpr int MAX_FRAME_SCORE{MAX_PINS};
+constexpr size_t MAX_PINS{10};
+constexpr size_t MAX_FRAMES{10};
+constexpr size_t MAX_ROLLS{21};
+constexpr size_t MAX_FRAME_SCORE{MAX_PINS};
 
 class Game {
 public:
-    void roll(int pins);
-    int score();
+    void roll(size_t pins);
+    size_t score();
     void reset();
 
 private:
-    int rolls[MAX_ROLLS]{};
-    int currentRoll{};
+    size_t rolls[MAX_ROLLS]{};
+    size_t currentRoll{};
 
-    bool isSpare(int firstFrameThrow);
-    bool isStrike(int firstFrameThrow);
-    int nextTwoBallsForStrike(int firstFrameThrow);
-    int nextBallForSpare(int firstFrameThrow);
-    int twoBallsInFrame(int firstFrameThrow);
+    bool isSpare(size_t firstFrameThrow);
+    bool isStrike(size_t firstFrameThrow);
+    size_t nextTwoBallsForStrike(size_t firstFrameThrow);
+    size_t nextBallForSpare(size_t firstFrameThrow);
+    size_t twoBallsInFrame(size_t firstFrameThrow);
 };
